third_party/WebKit/LayoutTests/fast/dom/shadow/tabstop-focus-highlight-sibling-expected.txt

crbug/474687 :focus style should properly be applied to shadow hosts.
PASS document.activeElement.id is "XMenu1"
PASS backgroundColorOf('XMenu1') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu2') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu3') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu4') is "rgb(170, 170, 170)"
PASS document.activeElement.id is "XMenu2"
PASS backgroundColorOf('XMenu1') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu2') is "rgb(204, 204, 204)"
PASS backgroundColorOf('XMenu3') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu4') is "rgb(170, 170, 170)"
PASS document.activeElement.id is "XMenu3"
PASS backgroundColorOf('XMenu1') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu2') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu3') is "rgb(204, 204, 204)"
PASS backgroundColorOf('XMenu4') is "rgb(170, 170, 170)"
PASS document.activeElement.id is "XMenu4"
PASS backgroundColorOf('XMenu1') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu2') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu3') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu4') is "rgb(204, 204, 204)"
PASS document.activeElement.id is "XMenu3"
PASS backgroundColorOf('XMenu1') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu2') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu3') is "rgb(204, 204, 204)"
PASS backgroundColorOf('XMenu4') is "rgb(170, 170, 170)"
PASS document.activeElement.id is "XMenu2"
PASS backgroundColorOf('XMenu1') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu2') is "rgb(204, 204, 204)"
PASS backgroundColorOf('XMenu3') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu4') is "rgb(170, 170, 170)"
PASS document.activeElement.id is "XMenu1"
PASS backgroundColorOf('XMenu1') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu2') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu3') is "rgb(170, 170, 170)"
PASS backgroundColorOf('XMenu4') is "rgb(170, 170, 170)"
PASS successfullyParsed is true

TEST COMPLETE
 
